pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memdio"
version = "0.2.0"
description = "AI memory encoded in sound. Lossless, fast, and hardware-adaptive."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Azhar", email = "azhar@example.com" }
]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "soundfile>=0.12.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "reedsolo>=1.7.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
search = [
    "fastembed>=0.7.0",
    "sqliteai-vector>=0.9.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
benchmark = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/azhar0406/memdio"
Repository = "https://github.com/azhar0406/memdio"

[project.scripts]
memdio = "memdio.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["memdio"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
